♻️ XJETF6 runner: extract blueprint plan helpers
