+++ .agentplane/tasks/202605282000-XJETF6/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605282000-XJETF6/task-runner-execution-usecase-decomposition",
+  "created_at": "2026-05-28T20:01:14.372Z",
+  "diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "last_verified_at": "2026-05-28T20:04:39.875Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605282000-XJETF6",
+  "updated_at": "2026-05-28T20:01:14.372Z",
+  "verify": {
+    "status": "pass"
+  }
+}
