+++ .agentplane/tasks/202605282000-XJETF6/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T20:01:14.372Z
+
+## Task
+
+- Task: `202605282000-XJETF6`
+- Title: Task runner execution usecase decomposition
+- Status: DOING
+- Branch: `task/202605282000-XJETF6/task-runner-execution-usecase-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282000-XJETF6/README.md`
+
+## Verification
+
+- State: ok
+- Note: Task runner usecase decomposition verified: task-run.ts reduced from 566 to 388 lines, blueprint plan/snapshot helpers extracted, targeted runner blueprint/lifecycle tests passed, typecheck passed, lint:core passed, format:changed passed, hotspot threshold check passed.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T20:01:14.372Z
+- Branch: task/202605282000-XJETF6/task-runner-execution-usecase-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
